fixed a bunch of problems incl renaming buckets + blob path, not showing proper data, etc

+++ routes/blobs.js
@@ -15,9 +15,12 @@ const storage = multer.diskStorage({
 
     try {
       const bucket = await Bucket.findById(id);
-      const pathName = path.join('/opt/workspace/myS3/', uuid, bucket.name);
-      cb(null, pathName);
-    } catch (e) {
+
+      if (bucket) {
+        const pathName = path.join('/opt/workspace/myS3/', uuid, id);
+        cb(null, pathName);
+      }
+    } catch (err) {
       throw new Error('could not get bucket');
     }
   },
@@ -40,52 +43,54 @@ api.post('/', upload.single('blob'), async (req, res) => {
     const { originalname, size } = req.file;
 
     const bucket = await Bucket.findById(id);
+    if (bucket) {
+      const pathName = path.join('/opt/workspace/myS3/', uuid, bucket.id.toString(), originalname);
 
-    const pathName = path.join('/opt/workspace/myS3/', uuid, bucket.name, originalname);
+      const blob = new Blob({
+        name: originalname,
+        path: pathName,
+        size,
+        bucket_id: id,
+      });
 
-    const blob = new Blob({
-      name: originalname,
-      path: pathName,
-      size,
-      bucket_id: id,
-    });
+      await blob.save();
 
-    await blob.save();
-
-    res.status(201).json({ data: { blob }, meta: {} });
+      res.status(201).json({ data: { blob }, meta: {} });
+    }
   } catch (err) {
-    res.status(400).json({ err });
+    res.status(400).json({ err: `could not create blob, err: ${err.message}` });
   }
 });
 
 api.post('/duplicate/:id', async (req, res) => {
-  const { id } = req.params;
-
   try {
+    const { id } = req.params;
     const blob = await Blob.findById(id);
-    const {
-      name, path, size, bucket_id,
-    } = blob;
 
-    const { name: filename, extension } = getExtension(name);
-    const newName = `${filename}.copy${extension}`;
-    let newPath = path.substr(0, path.length - (filename.length + extension.length));
-    newPath = `${newPath}${newName}`;
+    if (blob) {
+      const {
+        name, path, size, bucket_id,
+      } = blob;
 
-    const newBlob = new Blob({
-      name: newName,
-      path: newPath,
-      size,
-      bucket_id,
-    });
+      const { name: filename, extension } = getExtension(name);
+      const newName = `${filename}.copy${extension}`;
+      let newPath = path.substr(0, path.length - (filename.length + extension.length));
+      newPath = `${newPath}${newName}`;
 
-    Filesystem.duplicateBlob(path, newPath);
+      const newBlob = new Blob({
+        name: newName,
+        path: newPath,
+        size,
+        bucket_id,
+      });
 
-    await newBlob.save();
+      Filesystem.duplicateBlob(path, newPath);
+      await newBlob.save();
 
-    res.status(201).json({ data: { blob }, meta: {} });
+      res.status(201).json({ data: { newBlob }, meta: {} });
+    }
   } catch (err) {
-    res.status(400).json({ err: 'error' });
+    res.status(400).json({ err: `could not duplicate blob, err: ${err.message}` });
   }
 });
 
@@ -95,7 +100,7 @@ api.get('/', async (req, res) => {
     const blobs = await Blob.findAll({ where: { bucket_id: id } });
     res.status(200).json({ data: { blobs }, meta: {} });
   } catch (err) {
-    res.status(400).json({ err: `could not connect to database, err: ${err.message}` });
+    res.status(400).json({ err: `could not get blobs, err: ${err.message}` });
   }
 });
 
@@ -103,9 +108,12 @@ api.get('/metadata/:id', async (req, res) => {
   try {
     const { id } = req.params;
     const blob = await Blob.findById(id);
-    res.status(200).json({ data: { blob: { metadata: { size: blob.size, path: blob.path } } } });
+
+    if (blob) {
+      res.status(200).json({ data: { blob: { metadata: { size: blob.size, path: blob.path } } } });
+    }
   } catch (err) {
-    res.status(400).json({ err: `could not connect to database, err: ${err.message}` });
+    res.status(400).json({ err: `could not get blob metadata, err: ${err.message}` });
   }
 });
 
@@ -115,7 +123,7 @@ api.get('/:id', async (req, res) => {
     const blob = await Blob.findById(id);
     res.status(200).json({ data: { blob } });
   } catch (err) {
-    res.status(400).json({ err: `could not connect to database, err: ${err.message}` });
+    res.status(400).json({ err: `could not get blob, err: ${err.message}` });
   }
 });
 
@@ -123,13 +131,16 @@ api.delete('/:id', async (req, res) => {
   const { id } = req.params;
   try {
     const blob = await Blob.findById(id);
-    const { path } = blob;
-    await Blob.destroy({ where: { id } });
-    Filesystem.removeBlob(path);
 
-    res.status(204).json();
+    if (blob) {
+      const { path } = blob;
+      await Blob.destroy({ where: { id } });
+      Filesystem.removeBlob(path);
+
+      res.status(204).json();
+    }
   } catch (err) {
-    res.status(400).json({ err: `could not connect to database, err: ${err.message}` });
+    res.status(400).json({ err: `could not delete blob, err: ${err.message}` });
   }
 });
 
@@ -138,11 +149,14 @@ api.get('/download/:id', async (req, res) => {
 
   try {
     const blob = await Blob.findById(id);
-    const { path } = blob;
 
-    res.download(path);
-  } catch (e) {
-    res.status(400).json({ err: 'could not download file' });
+    if (blob) {
+      const { path } = blob;
+
+      res.download(path);
+    }
+  } catch (err) {
+    res.status(400).json({ err: `could not retrieve blob, err: ${err.message}` });
   }
 });
