+++ routes/buckets.js
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 import { pick } from 'lodash';
 import Bucket from '../models/bucket';
+import Blob from '../models/blob';
 import Filesystem from '../lib/filesystem';
 
 const api = Router();
@@ -14,13 +15,14 @@ api.post('/', async (req, res) => {
       name,
       user_uuid: uuid,
     });
+    await bucket.save();
 
     try {
-      Filesystem.createBucket(uuid, name);
-      await bucket.save();
+      Filesystem.createBucket(uuid, bucket.id);
+
       res.status(201).json({ data: { bucket }, meta: {} });
     } catch (err) {
-      res.status(400).json({ err: err.message });
+      res.status(400).json({ err: `could not create bucket, err: ${err.message}` });
     }
   } catch (err) {
     res.status(400).json({ err });
@@ -29,10 +31,12 @@ api.post('/', async (req, res) => {
 
 api.get('/', async (req, res) => {
   try {
-    const buckets = await Bucket.findAll();
+    const { uuid } = req.user;
+    const buckets = await Bucket.findAll({ where: { user_uuid: uuid } });
+
     res.status(200).json({ data: { buckets }, meta: {} });
   } catch (err) {
-    res.status(400).json({ err: `could not connect to database, err: ${err.message}` });
+    res.status(400).json({ err: `could not get buckets, err: ${err.message}` });
   }
 });
 
@@ -49,7 +53,7 @@ api.head('/:id', async (req, res) => {
       res.status(400).end();
     }
   } catch (err) {
-    res.status(400).json({ err: `could not connect to database, err: ${err.message}` });
+    res.status(400).json({ err: `could not get bucket, err: ${err.message}` });
   }
 });
 
@@ -57,9 +61,13 @@ api.get('/:id', async (req, res) => {
   try {
     const { id } = req.params;
     const bucket = await Bucket.findById(id);
-    res.status(200).json({ data: { bucket } });
+    const blobs = await Blob.findAll({ where: { bucket_id: id } });
+
+    if ((blobs, bucket)) {
+      res.status(200).json({ data: { bucket, blobs } });
+    }
   } catch (err) {
-    res.status(400).json({ err: `could not connect to database, err: ${err.message}` });
+    res.status(400).json({ err: `could get bucket, err: ${err.message}` });
   }
 });
 
@@ -69,34 +77,32 @@ api.delete('/:id', async (req, res) => {
     const { uuid } = req.user;
 
     const bucket = await Bucket.findById(id);
-    const { name } = bucket;
-    await Bucket.destroy({ where: { id } });
-    Filesystem.removeBucket(uuid, name);
-    res.status(204).json();
+
+    if (bucket) {
+      await Bucket.destroy({ where: { id } });
+      Filesystem.removeBucket(uuid, id);
+
+      res.status(204).json();
+    }
   } catch (err) {
-    res.status(400).json({ err: err.message });
+    res.status(400).json({ err: `could not delete bucket, err: ${err.message}` });
   }
 });
 
 api.put('/:id', async (req, res) => {
   try {
     const { id } = req.params;
-    const { uuid } = req.user;
 
     const bucket = await Bucket.findById(id);
-    const { name } = bucket;
 
     if (bucket) {
       const fields = pick(req.body, ['name']);
-      const newName = fields.name;
-
-      Filesystem.renameBucket(uuid, name, newName);
-
       await bucket.update(fields);
+
       res.status(204).json();
     }
   } catch (err) {
-    res.status(400).json({ err: err.message });
+    res.status(400).json({ err: `could not rename bucket, err: ${err.message}` });
   }
 });
